+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ All notable changes to this project are documented in this file.
 - Update NodeLeader peer monitoring system
 - Add ability to configure size of requests for blocks as well as block processing queue size
 - Update mainnet bootstrap files
+- Fix size calculations for all serializable classes
+- Add ``size`` key to JSON output of Block and Transaction
 
 [0.7.3] 2018-07-12
 -----------------------